# feat(csharp/src/Apache.Arrow.Adbc): improved performance of ValueAt helper and AdbcDataReader