--- a/csharp/src/Client/AdbcDataReader.cs
+++ b/csharp/src/Client/AdbcDataReader.cs
@@ -95,7 +95,7 @@ namespace Apache.Arrow.Adbc.Client
 
         public override object this[int ordinal] => GetValue(ordinal);
 
-        public override object this[string name] => GetOneValue(name);
+        public override object this[string name] => GetValue(GetOrdinal(name));
 
         public override int Depth => 0;
 
@@ -248,7 +248,15 @@ namespace Apache.Arrow.Adbc.Client
 
         public override int GetOrdinal(string name)
         {
-            return this.schema.GetFieldIndex(name);
+            if (this.columnNameToIndex == null)
+            {
+                this.columnNameToIndex = new Dictionary<string, int>(this.schema.FieldsList.Count);
+                for (int i = 0; i < this.schema.FieldsList.Count; i++)
+                {
+                    this.columnNameToIndex[this.schema.FieldsList[i].Name] = i;
+                }
+            }
+            return this.columnNameToIndex[name];
         }
 
         public override string GetString(int ordinal)
@@ -258,7 +266,13 @@ namespace Apache.Arrow.Adbc.Client
 
         public override object GetValue(int ordinal)
         {
-            object value = GetOneValue(ordinal);
+            IArrowArray arrowArray = this.RecordBatch.Column(ordinal);
+
+            // if the OnGetValue event is set, call it
+            object? value = OnGetValue?.Invoke(arrowArray, this.currentRowInRecordBatch);
+
+            // if the value is null, try to get the value from the ArrowArray
+            value = value ?? this.converters[ordinal](arrowArray, this.currentRowInRecordBatch) ?? DBNull.Value;
 
             if (value is SqlDecimal dValue && this.DecimalBehavior == DecimalBehavior.OverflowDecimalAsString)
             {
@@ -363,38 +377,6 @@ namespace Apache.Arrow.Adbc.Client
             }
 
             return dbColumns.AsReadOnly();
-        }
-
-        /// <summary>
-        /// Gets the value from a given column at the current row index.
-        /// </summary>
-        private object GetOneValue(string columnName)
-        {
-            if (this.columnNameToIndex == null)
-            {
-                this.columnNameToIndex = new Dictionary<string, int>(this.schema.FieldsList.Count);
-                for (int i = 0; i < this.schema.FieldsList.Count; i++)
-                {
-                    this.columnNameToIndex[this.schema.FieldsList[i].Name] = i;
-                }
-            }
-            return GetOneValue(this.columnNameToIndex[columnName]);
-        }
-
-        /// <summary>
-        /// Gets the value from a given column at the current row index.
-        /// </summary>
-        private object GetOneValue(int column)
-        {
-            IArrowArray arrowArray = this.RecordBatch.Column(column);
-
-            // if the OnGetValue event is set, call it
-            object? result = OnGetValue?.Invoke(arrowArray, this.currentRowInRecordBatch);
-
-            // if the value is null, try to get the value from the ArrowArray
-            result = result ?? this.converters[column](arrowArray, this.currentRowInRecordBatch) ?? DBNull.Value;
-
-            return result;
         }
 
         /// <summary>