--- a/csharp/test/Apache.Arrow.Adbc.Tests/DuckDbFixture.cs
+++ b/csharp/test/Apache.Arrow.Adbc.Tests/DuckDbFixture.cs
@@ -38,7 +38,7 @@ namespace Apache.Arrow.Adbc.Tests
 
             _databases = new Dictionary<string, AdbcDatabase>(StringComparer.OrdinalIgnoreCase);
 
-            string root = Directory.GetCurrentDirectory();
+            string root = Path.GetDirectoryName(typeof(DuckDbFixture).Assembly.Location) ?? throw new InvalidOperationException("directory");
             string file;
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
                 file = Path.Combine(root, "libduckdb.so");
@@ -47,6 +47,7 @@ namespace Apache.Arrow.Adbc.Tests
             else
                 file = Path.Combine(root, "libduckdb.dylib");
 
+            System.Console.WriteLine(file);
             _driver = CAdbcDriverImporter.Load(file, "duckdb_adbc_init");
         }